Several recipe tweaks.
Changes recipes that use black dye to use the tag instead of the specific item, adding compatibility with certain mods. This should be done with other dye colours as well, but was specifically added in this case for Dark Prismarine.

Simplifies silver block crafting to use shaped recipe (although this may be better as a replacement too).

Changes the backpack recipe back to wooden chests only - this previously had a typo so wasn't working, then was changed to all chests, but we don't want Mekanism's Personal Chest to be allowed in the craft (players will lose items).